Delete AssemblerAnchorsFromOverlapEvents.cpp and the tests that only test it
5026 lines of anchor-creation strategies with no production caller. Verified at
the linker rather than by grep: all six Assembler::createAnchorsFrom* symbols in
the object file are referenced by zero other objects in the executable.

That check alone was not enough, and nearly caused a bad deletion. Removing the
file broke the build on tests/test_integration.cpp, which used the six entry
points AND six dinara::testing::*ForTesting wrappers exposing the file's
vertex-splitting internals. The symbol scan had only covered the executable's
objects, not the test binary's -- so "no callers" was true of production and
false overall.

With that established the picture is closed: the splitting algorithms
(splitVertexByOverlapSupport, ...WithCoreMask, mclCluster, autoSplitVertex,
splitVertexByCliqueCover, peelGroups, ...WithPhasing) have no reference outside
this file either. The whole unit is reachable only from tests of itself, so the
tests go with it -- 14 TEST_CASEs, 8 of them testing the splitting internals
through the ForTesting wrappers.

    src/AssemblerAnchorsFromOverlapEvents.cpp   5026 lines, deleted
    src/Assembler.hpp                             56 declaration lines
    tests/test_integration.cpp                  1864 -> 1312 lines

Verified: clean full rebuild with zero errors, test suite passes (3760
assertions in 81 test cases, down from 95), and the fixture is byte-for-byte
unaffected -- 26695 anchors exported, 420 het, 576645 members, min 2, 0 empty,
103214 edges with 0 mismatches / 0 missing / 0 disabled, round trip matching.

Also clears the four compiler warnings in files this work touched: an
enum/uint8_t conditional in HifiasmImportedCigarStore.hpp, a sign-compare on
--Kmers.k, and two set-but-unused loop counters in main.cpp. Warnings remain in
Shasta2AnchorGraph.cpp and the AnchorWindows files, which this did not touch.

+++ src/HifiasmImportedCigarStore.hpp
@@ -121,8 +121,8 @@ namespace dinara {
                 const CigarToken raw(tokens[i]);
                 const uint8_t op = raw.op();
                 const uint8_t dinaraOp =
-                    (op == CigarOpIns) ? CigarOpDel :
-                    (op == CigarOpDel) ? CigarOpIns : op;
+                    (op == CigarOpIns) ? uint8_t(CigarOpDel) :
+                    (op == CigarOpDel) ? uint8_t(CigarOpIns) : op;
                 arena.emplace_back(CigarToken(dinaraOp, raw.len()));
                 if(opConsumesQuery(dinaraOp))  querySpan  += raw.len();
                 if(opConsumesTarget(dinaraOp)) targetSpan += raw.len();
